Add find_*_before functions & utils dir
Added both the find_first_before and find_last_before functions.
find_first_before finds the first element in a range for which the first
of two unary predicates returns true while ensuring that this element is
before a second element which causes the second undary predicate to
return true. If the first unary predicate never returns true, then the
function returns the element for which the second predicate returns
true, or the last element in the range, if the second predicate never
returns true.

find_last_before is an analogous function which returns the last element
for which a unary predicate returns true while being bounded by the
first element that causes the second predicate to return true & the end
of the range.

TODO: More tests need to be written. Basic verification achieved, but
more thorough testing must be done.